+++ Modules/_testcapi/parts.h
@@ -64,5 +64,6 @@ int _PyTestCapi_Init_Config(PyObject *mod);
 int _PyTestCapi_Init_Import(PyObject *mod);
 int _PyTestCapi_Init_Frame(PyObject *mod);
 int _PyTestCapi_Init_Type(PyObject *mod);
+int _PyTestCapi_Init_Function(PyObject *mod);
 
 #endif // Py_TESTCAPI_PARTS_H
