+++ PCbuild/_testcapi.vcxproj.filters
@@ -123,6 +123,9 @@
     <ClCompile Include="..\Modules\_testcapi\type.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Modules\_testcapi\function.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc">
